Added text domain and resolved editor issue
