+++ assets/css/blocks/fse-post-author.css
@@ -1,0 +1,1 @@
+.uagb-post-author__wrap{display:inherit}.uagb-post-author__wrap .dashicons-admin-users.dashicons{vertical-align:inherit}
