+++ assets/css/blocks/fse-post-date.css
@@ -1,0 +1,1 @@
+.uagb-post-date__wrap>time>span{font-size:25.5px;vertical-align:initial;display:inline}
