+++ assets/css/blocks/fse-post-excerpt.css
@@ -1,0 +1,1 @@
+.uagb-post-comments__avatar-wrap{margin-left:10px;margin-right:10px;margin-top:5px;margin-bottom:5px}.uagb-post-comments__author{border-bottom:1px solid #e8e8e8}.uagb-post-comments__wrap .uagb-post-comments__content p{margin:0}
