+++ src/blocks.js
@@ -40,10 +40,10 @@ import "./blocks/wp-search/block.js"
 import "./blocks/review/block.js"
 import "./blocks/taxonomy-list/block.js"
 import "./blocks/lottie/block.js"
-import "./blocks/post-author/block.js"
-import "./blocks/post-date/block.js"
-import "./blocks/post-categories/block.js"
-import "./blocks/post-excerpt/block.js"
+import "./blocks/fse-post-author/block.js"
+import "./blocks/fse-post-date/block.js"
+import "./blocks/fse-post-categories/block.js"
+import "./blocks/fse-post-excerpt/block.js"
 
 import UAGB_Block_Icons from "../dist/blocks/uagb-controls/block-icons"
