+++ src/blocks/fse-post-author/style.scss
@@ -1,0 +1,6 @@
+.uagb-post-author__wrap {
+    display: inherit;
+    .dashicons-admin-users.dashicons{
+        vertical-align: inherit;
+    }
+}
