+++ src/blocks/fse-post-date/style.scss
@@ -1,0 +1,5 @@
+.uagb-post-date__wrap > time > span {
+    font-size: 25.5px;
+    vertical-align: initial;
+    display: inline;
+}
