--- src/blocks/post-author/style.scss
@@ -1,6 +1,0 @@
-.uagb-post-author__wrap {
-    display: inherit;
-    .dashicons-admin-users.dashicons{
-        vertical-align: inherit;
-    }
-}
