--- src/blocks/post-date/editor.scss
@@ -1,5 +1,0 @@
-.uagb-post-date__wrap > time > span {
-    font-size: 25.5px;
-    vertical-align: initial;
-    display: inline;
-}
