--- src/blocks/post-date/style.scss
@@ -1,4 +1,0 @@
-.uagb-post-date__wrap > time > span {
-    vertical-align: initial;
-    display: inline;
-}
